user friendly logging - INFO is default level, add -q and progressbar
Previously the logdetective does not print anything for long time and seems to be stuck.
I changed default level to INFO.
I added code to have meaningfull use of -v and not require -vvvvvvvvvv to get to INFO level.
And I added progressbar when downloading the model.
For integration I added --quite that surpress logging and only print result.