Loosen build_name regex
The version of this we receive from the CI_JOB_NAME_SLUG and the one we
receive from job_event webhooks differs. (One uses underscores, the
other uses hyphens). Since we really only care that it begins and ends
with "build" and "rpm", respectively, we'll just ignore the formatting
in the middle entirely.

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>