Generate Gitlab merge request comments
This patch adds support for converting the results of the staged
analysis into a comment to be submitted against the original merge
request that triggered the analysis.

It takes advantage of Jinja 2 templating to generate the content in
Markdown format.

This patch also modifies the return values of the log preprocessing
function to return the URL of the detected log so that it can be linked
in the comment.

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>