Separate comment template into short and full
In order to not flood users with long emails, we will send only the
summary out in the initial comment and then submit an update after
five seconds. This works because Gitlab doesn't send an email to
notify about edits to comments, only the initial comment creation.

It includes a five-second sleep in order to ensure that the
creation and edit don't get bundled together closely enough that
they go out as a single email.

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>