Separate reactor for Gitlab into its own process
This is a major refactor of code that creates a new logdetective.reactor
module in the project. It will act as its own service to listen for
requests from Gitlab for RPM builds from the RHEL-on-Gitlab pipeline,
process them and transmit the log of build failures to the main Log
Detective process.

It also switches out the use of python-requests for python-aiohttp
when performing GET and POST requests in the reactor as a first
step towards maintaining a native asynchronous framework.

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>