Load balance LLM queries to different endpoints
Fix LD-64

This breaks a backward compatibility of the config file (the `inference` field
now expect a list of configurations) but shouldn't break a previous behavior in
a case when the list contains just one inference configuration.

When there are multiple configurations, we improve our fault tolerance because
if the inference request fails, we try next inference configuration, and then
the next one, etc.