DB: handle very large integers
The job_id values already exceed signed 32-bit integer space, which is
what postgresql uses when the Integer type is chosen. We hadn't noticed
it earlier because sqlite (which we were using for tests) treats Integer
as a 64-bit signed integer type.

This makes all of the identifiers explicitly BigInteger (64-bit or
larger) to handle future compatibility. There should be no meaningful
performance impact unless someone is running on a 32-bit kernel.

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>