Track metrics for Gitlab-triggered requests
Gitlab uses the underlying code for staged analysis, but it doesn't call
the endpoint directly, meaning it wasn't being saved to the database.
This patch adds an argument to the decorator allowing us to override the
function name it uses to save to the DB (for backwards compatibility)
and moves the decorator to the inner function perform_staged_analysis().

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>