Separate GET and POST actions for Koji
This rewrites the API significantly, eliminating the need for a POST
body and making it possible to query for results without triggering a
new analysis.

Requests can now be made as a GET to
/analyze/rpmbuild/koji/{koji_instance}/{task_id} to just retrieve
results or a POST to the same URL to get-or-retrieve the results.

Both endpoints still require the X-Koji-Token header to be set
appropriately.

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>