Implement a callback mechanism for Koji analyses
Clients initiating an analysis request may optionally send the
"X-Koji-Callback" header, providing a URL to be notified when the
analysis completes.

Upon completion, Log Detective will send an HTTP POST request to this
callback address, with a simple JSON payload containing only the
completed task ID. The client will be responsible for sending a
subsequent GET request if they want to retrieve the full response.

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>