Snippet filtering by estimated relevance
Snippets can now be filtered by relevance score. The score is assigned by LLM
and returned as part of structured output in a JSON form.

Snippets with top-k score are kept and send for final analysis, the rest is discarded.
In the event that all snippets recieved the same score, no filtering occurs,
and all snippets are used instead.

All snippets are also returned if requested number of top snippets is equal,
or greater then number of snippets in original response.

Signed-off-by: Jiri Podivin <jpodivin@redhat.com>